+++ Clase/King.cpp
@@ -15,13 +15,13 @@ King::King(int x) {
     {
         texture.loadFromFile("Texturi/whiteking.png", sf::IntRect(0,0,188,309));
         sprite.setTexture(texture);
-        sprite.setScale(100.0/218, 100.0/339);
+        sprite.setScale(float(100.0/218), float(100.0/339));
     }
     else
     {
         texture.loadFromFile("Texturi/blackking.png", sf::IntRect(0,0,196,306));
         sprite.setTexture(texture);
-        sprite.setScale(100.0/226, 100.0/336);
+        sprite.setScale(float(100.0/226), float(100.0/336));
     }
 }
