+++ Clase/Queen.cpp
@@ -13,13 +13,13 @@ Queen::Queen(int x) {
     {
         texture.loadFromFile("Texturi/whitequeen.png", sf::IntRect(0,0,183,286));
         sprite.setTexture(texture);
-        sprite.setScale(100.0/200, 100.0/300);
+        sprite.setScale(float(100.0/200), float(100.0/300));
     }
     else
     {
         texture.loadFromFile("Texturi/blackqueen.png", sf::IntRect(0,0,182,276));
         sprite.setTexture(texture);
-        sprite.setScale(100.0/200, 100.0/300);
+        sprite.setScale(float(100.0/200), float(100.0/300));
 
     }
 }
